[IMP] project, _*:  improve the project update form view description
_*: sale_timesheet, sale_project

Currently, profitability isn't displayed in the project update description
form without the sale_timesheet module. In this commit, we have improved
the project description according the project side panel.

task-3916169
