+++ addons/sale_project/models/__init__.py
@@ -6,6 +6,7 @@ from . import project_milestone
 from . import project_project
 from . import project_task_recurrence
 from . import project_task
+from . import project_update
 from . import res_config_settings
 from . import sale_order_line
 from . import sale_order_template_line
