+++ addons/sale_timesheet/models/__init__.py
@@ -10,7 +10,6 @@ from . import product_template
 from . import project_project
 from . import project_sale_line_employee_map
 from . import project_task
-from . import project_update
 from . import res_config_settings
 from . import sale_order_line
 from . import sale_order
